Revert "Logica modelo interno hdu10"
