+++ src/main/java/com/app/SIGET/ws/SpringWebSocket.java
@@ -43,7 +43,9 @@ public class SpringWebSocket extends TextWebSocketHandler {
 		}
 		
 		if ("eliminar".equals(jso.getString(TYPE))) {
-			Manager.get().eliminarUsuario((String) jso.get(NOMBRE));
+			Manager.get().eliminar((String) jso.get(NOMBRE));
+			// session.sendMessage(new TextMessage(Manager.get().leer().toString()));
+
 		}
 	}
