--- src/test/java/features/eliminarUsuario.feature
@@ -1,25 +1,0 @@
-#Author: your.email@your.domain.com
-#Keywords Summary :
-#Feature: List of scenarios.
-#Scenario: Business rule through list of steps with arguments.
-#Given: Some precondition step
-#When: Some key actions
-#Then: To observe outcomes or validation
-#And,But: To enumerate more Given,When,Then steps
-#Scenario Outline: List of steps for data-driven as an Examples and "placeholder"
-#Examples: Container for s table
-#Background: List of steps run before each of the scenarios
-#""" (Doc Strings)
-#| (Data Tables)
-#@ (Tags/Labels):To group Scenarios
-#"" (placeholder)
-#""
-## (Comments)
-#Sample Feature Definition Template
-@tag
-Feature: Eliminar usuario
-
-  @tag1
-  Scenario: Eliminar usuario
-    Given En la vista admin ver usuarios seleccionar uno por "nombre" y eliminarlo
-    Then Se eliminar el usuario con nombre "nombre"
